Added visibility.  GameEngine initializes elements in Grid.
Room is given a briefcase in the constructor rather than through a
method.
Spy now holds its initial position and the visibility range.
VisibleMark overrides isVisible to always be true.
The VisiblePair nested class in Grid keeps track of GameObjects that are
visible.  It is added to by setAsVisible() and is cleared in
setToInvisible().
UserInterface now has a loop for player input.