Ninjas move and attack, powerups working, and UI improvements
Ninjas now move randomly (there is no way for it to become an infinite
loop but sometimes a Ninja may not be able to move).  Ninjas will attack
the Spy if it is in range.

The GameEngine.displayBoard() now shows the status of power ups and
ammunition.  This meant that Grid had to be printed out by row,
therefore Grid.printRow(int) was added.  It is super easy to add to this
print out on the side of the grid by adding to the String in
GameEngine.generateBoardInformation().

All PowerUps work properly.  Take note of the methods
GameEngine.useSpyPowerups() and GameEngine.updateSpyPowerups().
useSpyPowerups() handles using the GameObject.belowObject and is called
by the UserInterface before GameEngine.enemyTurn() this allows powerups
that were just picked up to apply to the next GameEngine.enemyTurn().
updateSpyPowerups() decrements Spy.invisibilityTurn and reveals the
briefcase if spy.hasRadar().  It is called by the UserInterface after
enemyTurn() so that in the Player's next turn they are informed that
their powerups have expired.

The Bullet powerup is taken care of too.  The powerup is only used if
the Spy needs ammo.  This was done by changing Useable.useOn() to return
a boolean (rather than void).  If it returns true, then it was used
successfully and can be set to null.  If it returns false, then it was
not used succesfully and should not be set to null.

Got rid of useless field GameEngine.rooms and put in the very usefull
field GameEngine.briefcaseRoom which stores the room with the briefcase
in it.  This is so when the  Spy.hasRadar(), we can just call
briefcaseRoom.revealBriefCase().

In the meeting we said, "The spy would never be able to step on a Ninja"
and "The ninja could never step on Spy."  That is incorrect.  If the Spy
is invincible it could try to step on the Ninja and a Ninja could
potentially try to step on the Spy.  Therefore, both of their overriden
GameObject.stepOn(DIRECTION) methods return MOVE_RESULT.ILLEGAL along
with an appropiate message.

I also got rid of the unused attack method in Weapon (its been replaced
by the two much better attack methods).