UI for save and load. Now have 3 attempts.  Reset works.
The UserInterface now has options to save and load.  However, these
options don't actually save or load yet as GameEngine.save() and
GameEngine.load() have not been implemented.

The player now has 3 attempts.  The Spy is set back to the initial state
upon every death and all Ninjas are cleared from that spawn area.

The game now fully resets properly (when you start a new game a second
time).

I don't think canSetNinja() was working correctly.  It was adding the
difference in the x and y and checking if that was greater than 3.
Instead, it should have been individually been checking that x and y
were both greater than 3.  I'll make an issue for this.