Can now load, save, and exit game
Had to implement Serializable in parent classes.  Not doing so would
result in a ClassNotFoundException when serializing.

Had to create readObject(ObjectInputStream) method in the Grid.  This is
called when deserializing the grid object and works like the normal
deserialization, but it will call setToInvisible().  It calls
setToInvisible because if the player saved in playerTurn(), the
visiblemarks would still be there and the player would have another
opportunity to look around.

Removed MOVE_STATUS.POWER_UP because we don't need it in this
implementation.  spy.usePowerups() is called every turn instead.