Fixed visiblity and removed recursion
First of all, I removed getUserCommand(), its not being used.  Secondly,
I removed gridString.  It is better that grid be up to date and we don't
have to save what i looked like.  Although I am not opposed to adding it
back in to save processing time.  However, that is not a priority.  The
visiblity is reset properly in game engine.